Fix 296 compile errors: JVM-only APIs, imports, concurrency, type constraints
Systematic cleanup reducing compile errors from 793 to 497:

- Replace JVM-only APIs with KMP alternatives: sortedMapOf, toSortedMap,
  String.format, Character.isValidCodePoint, putIfAbsent
- Add expect/actual for runBlocking (JVM/Native/JS/WasmJS)
- Add expect/actual for ReentrantLock (JVM/Native/JS/WasmJS)
- Fix atomicfu references to use kotlin.concurrent.atomics API
- Fix wrong import paths in ~20 files (non-existent sub-packages)
- Remove invalid where clauses on non-generic functions (7 files)
- Fix public inline functions accessing internal members (7 files)
- Add Rust Deref delegation methods to ArcTy (isAny/isNever)
- Rewrite ToValue.kt removing broken String.toValue
- Merge expect/actual SimdImpl into single non-expect function
- Fix fmtWithConfig call sites for String-returning variants
- Fix val reassignment in TypeAnnot tests (use copy())
- Fix regex DOT_MATCHES_ALL (KMP common only has 3 options)
- Fix various type mismatches (UInt/Int, Result unwrapping)
- Fix default() calls to use actual constructors/companions

+++ src/commonMain/kotlin/io/github/kotlinmania/starlark_kotlin/eval/bc/definitely_assigned/DefinitelyAssigned.kt
@@ -51,7 +51,7 @@ class BcDefinitelyAssigned private constructor(
     /** Is local variable definitely assigned at given program point? */
     // pub(crate) fn is_definitely_assigned(&self, local: LocalSlotId) -> bool
     internal fun isDefinitelyAssigned(local: LocalSlotId): Boolean {
-        return definitelyAssigned[local.index]
+        return definitelyAssigned[local.index.toInt()]
     }
 
     /**
@@ -67,7 +67,7 @@ class BcDefinitelyAssigned private constructor(
      */
     // pub(crate) fn mark_definitely_assigned(&mut self, local: LocalSlotId)
     internal fun markDefinitelyAssigned(local: LocalSlotId) {
-        definitelyAssigned[local.index] = true
+        definitelyAssigned[local.index.toInt()] = true
     }
 
     /**
